Renaming contracts to validation
