--- contracts/doc.go
@@ -1,2 +1,0 @@
-// Package contracts provides utilities for verifying Ethereum based smart contracts.
-package contracts
