+++ validation/doc.go
@@ -1,0 +1,2 @@
+// Package validation provides utilities for verifying Ethereum based smart contracts.
+package validation
